+++ appGlueLib/src/main/java/com/appglue/datatypes/IOType.java
@@ -4,7 +4,7 @@ import android.os.Bundle;
 
 public abstract class IOType
 {
-	protected int id;
+	protected long id;
 
 	protected String name;
 	protected String className;
@@ -23,7 +23,7 @@ public abstract class IOType
 		this.name = name;
 	}
 	
-	public IOType(int id, String name, String className)
+	public IOType(long id, String name, String className)
 	{
 		this.id = id;
 		this.name = name;
@@ -34,12 +34,12 @@ public abstract class IOType
         return this.id + ": " + this.name;
     }
 	
-	public int getID()
+	public long getID()
 	{
 		return this.id;
 	}
 
-	public void setID(int id)
+	public void setID(long id)
 	{
 		this.id = id;
 	}
